Hippique-Analyse-0925/requirements.txt
# --- Core scraping / parsing ---
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=5.2.1
tenacity>=8.3.0
tqdm>=4.66.4

# --- Data / fichiers ---
pandas>=2.2.2
numpy>=1.26.4
openpyxl>=3.1.2

# --- YAML / calibration EV ---
PyYAML>=6.0.1

# --- API / web ---
fastapi==0.112.2
uvicorn[standard]==0.30.6

# --- Google Cloud Storage (utilisé par drive_sync.py) ---
google-cloud-storage>=2.18.0
google-auth>=2.29.0

# --- Robustesse encodages ---
charset-normalizer>=3.3.2



# --- Tests ---
pytest>=7.4




